Move element history actions to old element controllers
