+++ app/views/browse/_version_actions.erb
@@ -21,7 +21,7 @@
     <% end %>
     &middot;
   <% end %>
-  <%= link_to t("browse.view_history"), :controller => :browse, :action => "#{@type}_history" %>
+  <%= link_to t("browse.view_history"), :action => :index %>
   <% unless @feature.latest_version? %>
     &middot;
     <%= link_to({ :version => @feature.version + 1 }, { :class => "icon-link" }) do %>
