+++ test/controllers/old_relations_controller_test.rb
@@ -3,9 +3,56 @@ require "test_helper"
 class OldRelationsControllerTest < ActionDispatch::IntegrationTest
   def test_routes
     assert_routing(
+      { :path => "/relation/1/history", :method => :get },
+      { :controller => "old_relations", :action => "index", :id => "1" }
+    )
+    assert_routing(
       { :path => "/relation/1/history/2", :method => :get },
       { :controller => "old_relations", :action => "show", :id => "1", :version => "2" }
     )
+  end
+
+  def test_history
+    relation = create(:relation, :with_history)
+    sidebar_browse_check :relation_history_path, relation.id, "browse/history"
+    assert_select "h4", /^Version/ do
+      assert_select "a[href='#{old_relation_path relation, 1}']", :text => "1", :count => 1
+    end
+  end
+
+  def test_history_of_redacted
+    relation = create(:relation, :with_history, :version => 4)
+    relation_v1 = relation.old_relations.find_by(:version => 1)
+    relation_v1.redact!(create(:redaction))
+    relation_v3 = relation.old_relations.find_by(:version => 3)
+    relation_v3.redact!(create(:redaction))
+
+    get relation_history_path(:id => relation)
+    assert_response :success
+    assert_template "browse/history"
+
+    # there are 4 revisions of the redacted relation, but only 2
+    # should be showing details here.
+    assert_select ".browse-section", 4
+    assert_select ".browse-section.browse-redacted", 2
+    assert_select ".browse-section.browse-relation", 2
+  end
+
+  def test_unredacted_history_of_redacted
+    session_for(create(:moderator_user))
+    relation = create(:relation, :with_history, :version => 4)
+    relation_v1 = relation.old_relations.find_by(:version => 1)
+    relation_v1.redact!(create(:redaction))
+    relation_v3 = relation.old_relations.find_by(:version => 3)
+    relation_v3.redact!(create(:redaction))
+
+    get relation_history_path(:id => relation, :params => { :show_redactions => true })
+    assert_response :success
+    assert_template "browse/history"
+
+    assert_select ".browse-section", 4
+    assert_select ".browse-section.browse-redacted", 0
+    assert_select ".browse-section.browse-relation", 4
   end
 
   def test_visible_with_one_version
